+++ expected/wasm32-wasip2/defined-symbols.txt
@@ -185,15 +185,14 @@ __progname
 __progname_full
 __pthread_cond_timedwait
 __pthread_create
-__pthread_detach
 __pthread_join
 __pthread_key_create
 __pthread_key_delete
-__pthread_mutex_consistent
 __pthread_mutex_lock
 __pthread_mutex_timedlock
 __pthread_mutex_trylock
 __pthread_mutex_unlock
+__pthread_once
 __pthread_rwlock_rdlock
 __pthread_rwlock_timedrdlock
 __pthread_rwlock_timedwrlock
@@ -1127,13 +1126,14 @@ pthread_condattr_setpshared
 pthread_create
 pthread_detach
 pthread_equal
-pthread_exit
+pthread_getattr_np
 pthread_getspecific
 pthread_join
 pthread_key_create
 pthread_key_delete
 pthread_mutex_consistent
 pthread_mutex_destroy
+pthread_mutex_getprioceiling
 pthread_mutex_init
 pthread_mutex_lock
 pthread_mutex_timedlock
@@ -1457,6 +1457,7 @@ tgamma
 tgammaf
 tgammal
 thrd_current
+thrd_detach
 thrd_equal
 thrd_sleep
 time
