+++ thread-stub/pthread_barrier_destroy.c
@@ -1,0 +1,6 @@
+#include "pthread_impl.h"
+
+int pthread_barrier_destroy(pthread_barrier_t *b)
+{
+	return 0;
+}
