+++ thread-stub/pthread_cond_init.c
@@ -1,0 +1,6 @@
+#include "pthread_impl.h"
+
+int pthread_cond_init(pthread_cond_t *restrict c, const pthread_condattr_t *restrict a)
+{
+	return 0;
+}
