+++ thread-stub/pthread_cond_signal.c
@@ -1,0 +1,6 @@
+#include "pthread_impl.h"
+
+int pthread_cond_signal(pthread_cond_t *c)
+{
+	return 0;
+}
